refactor : slider 파일 정리
