+++ client/src/components/Slider/Slider.css
