+++ client/src/components/Slider/Slider.js
